publishfeed/requirements_lambda.txt
requests
feedparser
boto3
tweepy
pyyaml
python-dateutil
beautifulsoup4
openai==0.28
numpy
pandas
lxml
lxml_html_clean
opengraph-py3
rapidfuzz
newspaper3k==0.2.8
